Bug fixes for updatedAt/updatedBy properties and latest status value
In updatePost it is a little confusing. The Post alwasy has createdAt/createdBy and has no notion of updatedAt/By
Mixing them up meant the denormalized contactRef in the campaign.contacts array was getting it's updatedAt/By fields blown away by the update.

This commit fixes that.

We have also been passing undefined instead of the campaign.contacts.$.status into a couple of components that display the status.

That is also fixed in this commit.